Improved component routing
